refactor(async-agents): extract task lifecycle helpers to separate module
Move task lifecycle helpers out of async-agents into task-lifecycle module for clearer separation.
